Add coverage summarizer script for Tenzor
This script walks a coverage-instrumented build directory, runs gcov in JSON mode on each .gcda file, aggregates per-source line coverage, and generates a concise report. The report includes overall line coverage percentage, per-subsystem breakdown, top files by uncovered lines, and files that were compiled but never executed. A JSON dump of the per-file data is also created for downstream tooling.

+++ include/tenzor/autograd/ops.hpp
@@ -498,7 +498,7 @@ auto selu(const Variable& input) -> Variable;
 auto mish(const Variable& input) -> Variable;
 
 /// Leaky ReLU with negative slope. Grad: x>0 ? 1 : negative_slope
-auto leaky_relu(const Variable& input, float negative_slope = 0.01f) -> Variable;
+auto leaky_relu(const Variable& input, double negative_slope = 0.01) -> Variable;
 
 /// Softplus activation: log(1 + exp(β*x))/β. Grad: σ(β*x)
 auto softplus(const Variable& input, float beta = 1.0f) -> Variable;
